Extract ResearchCoinIcon into its own file
